Added question counter
Small thing: when I first started playing, I had no idea how many questions it would ask before I was done with the game.  That was a little bit frustrating/disorienting.  So this adds a line that prints out "Question X of Y" for every question.  It's dynamic and will change if you ever change the length of question_set.